Add public variables to barton recipe
This commit:
1. Adds various bitbake variables to serve as facade knobs for enabling
barton options. This will abstract DEPENDS and RDEPENDS management as
a convenience for clients.
2. Adds a barton_%.bbappend for the barton-example-app that turns on
matter.
3. Adds a default BBMASK entry for barton-example-app so that
its bbappend does not conflict with any other distribution
bbappend entry. Developers may comment out the BBMASK to build
the barton-example-app.

This is a prescursor to being able to build the reference app as the
barton-example-app.