Add Release Actions (#28)
This commit adds actions to start and finish a release via the git-flow
workflow. They are meant to be run manually by a maintainer. Start will
generate the release branch for a release, allowing manual release prep
on that branch. Finish will merge the branch to main, tag the release,
merge the tag back up into develop, and then delete the release branch.