Convert barton-example-app to build barton reference app (#21)
Until now, we have had a small, test program acting as the
barton-example-app, despite barton shipping with a reference application
that should be the go-to for devs. This was because building the
reference app required a newer version of linenoise than what was
available in relevant distributions (eg. RDK banana pi 4 reference
device). This commit:

1. Adds barton-linenoise package which provides the newer version of
linenoise. This is a stopgap until relevant distributions upgrade. No
new recipe should use this, including clients of barton. We reserve the
right to delete this recipe at any time.

2. Added a BARTON_BUILD_REFERENCE interface for barton recipe. Setting
this to ON builds the reference app in the barton recipe itself. This
commit also adds install instructions for the reference app binary.

3. Delete previous barton-example-app code and modifies recipe to also
download BartonCore but build the reference app in isolation (note:
separate from setting BARTON_BUILD_REFERENCE).

4. Fix some variable appends bugs in barton_1.1.0 and barton 2.0.0
recipes. These would manafest in missing spaces if multiple options were
selected.