remove `files/` requirement from generate_zzz (#36)
The generate_zzz.sh script initially required that the zap file passed
into the script must be in a `*files/` directory. This requirement is
not necessary and creates complications when a client does not want
their zap file within a `files/` dir.

Remove this requirement to make script execution simpler.